requirements.txt
asgiref==3.7.2
cssselect==1.2.0
Django==5.0.1
django-cors-headers==4.3.1
django-countries==7.5.1
django-extensions==3.2.3
django-filter==23.5
django-jsignature==0.12
django-seed==0.3.1
djangorestframework==3.14.0
drf-extra-fields==3.7.0
Faker==21.0.0
filetype==1.2.0
gunicorn==21.2.0
joblib==1.3.2
lxml==5.1.0
Markdown==3.5.1
mysqlclient==2.2.1
numpy==1.26.3
packaging==23.2
pandas==2.2.0
Pillow==10.1.0
psycopg==3.1.17
psycopg2==2.9.9
pyquery==2.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.3.0
toposort==1.10
typing_extensions==4.9.0
tzdata==2023.3
wheel==0.41.2
